requirements.txt
git+https://github.com/Rapptz/discord.py@master
transliterate==1.10.2
pillow==9.2.0
requests==2.28.1
emoji==2.0.0
motor==3.0.0
dnspython==2.2.1
youtube-transcript-api==0.4.4
lavalink==3.1.7
beautifulsoup4==4.11.1
python-magic==0.4.27
lxml==4.9.1
